cloud functions/extract-text-from-image/requirements.txt
google-cloud-vision>=3.4.1
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.8.0